extract representation logic into the representer
